refactor: Update /get-user route to decode JWT token and return user information
This commit updates the /get-user route in authRoutes.ts to decode the JWT token from the Authorization header and return the user information. It adds the decodeToken function from jwt.ts to decode the token and extract the user ID and username. If the token is invalid, it returns a 401 status with an error message. This change improves the security and functionality of the protected route.