feat: Add 'submitted' and 'verified' fields to transformed jobs
The code changes in this commit modify the 'getAssignedJobs' function in 'jobController.ts'. Two new fields, 'submitted' and 'verified', are added to the transformed jobs data. The 'submitted' field is set to true if the job has a submission time, and false otherwise. The 'verified' field is taken from the 'JobsUsers' object. This enhancement improves the information available for each job in the response.